Refactored UtilsScript.cs
Moved all action updating and responses to a new script. UtilsScript.cs is now almost entirely direct input response updates.
Removed Config.cs's pause flag and replaced most instances with UtilsScript.cs's InputStopped flag.
