+++ Assets/Scripts/GamePlay/StartGame.cs
@@ -36,9 +36,6 @@ public class StartGame : MonoBehaviour
 
     private void Start()
     {
-        // pause the game until the game is loaded and the transition is complete
-        Config.Instance.gamePaused = true;
-
         // load the game to the point that it can be played
         if (!GameLoader.Instance.LoadGame())
         {
@@ -87,11 +84,11 @@ public class StartGame : MonoBehaviour
     }
 
     private IEnumerator FadeGameplayIn()
-    {        
+    {
         Image fadeInScreen = this.gameObject.GetComponent<Image>();
         fadeInScreen.enabled = true;
         yield return Animate.FadeImage(fadeInScreen, GameValues.FadeColors.blackFadeOut, GameValues.AnimationDurataions.startGameFadeIn);
         fadeInScreen.enabled = false;
-        Config.Instance.gamePaused = false;
+        UtilsScript.Instance.InputStopped = false;
     }
 }
