+++ Assets/Scripts/UIScripts/ButtonScript.cs
@@ -35,9 +35,9 @@ public class ButtonScript : MonoBehaviour
     public void PauseGameButton()
     {
         Debug.Log("UI Button pause game");
-        if (UtilsScript.Instance.InputStopped || Config.Instance.gamePaused) return;
+        if (UtilsScript.Instance.InputStopped) return;
+        UtilsScript.Instance.InputStopped = true;
         Time.timeScale = 0;
-        Config.Instance.gamePaused = true;
         SoundEffectsController.Instance.PauseMenuButtonSound();
         MusicController.Instance.Paused = true;
         SceneManager.LoadScene(Constants.ScenesNames.pause, LoadSceneMode.Additive);
@@ -48,8 +48,8 @@ public class ButtonScript : MonoBehaviour
         Debug.Log("UI Button resume game");
         Time.timeScale = 1;
         SceneManager.UnloadSceneAsync(Constants.ScenesNames.pause);
-        Config.Instance.gamePaused = false;
         MusicController.Instance.Paused = false;
+        UtilsScript.Instance.InputStopped = false;
     }
 
     public void RestartGameButton()
@@ -66,6 +66,7 @@ public class ButtonScript : MonoBehaviour
         {
             MusicController.Instance.Paused = false;
         }
+        UtilsScript.Instance.InputStopped = false;
     }
 
     public void PlayAgain()
